refactor(features/partyroom): useUnregisterMeFromQueue mutation 승격 (widget → feature)
suppressNextSelfDjDeregister side-effect 보존. 데스크탑 통합 테스트 (dj-queue.integration.test.ts)
는 widget 위치 유지하고 import path 만 갱신 (spec §11.3 잠금).
