+++ src/features/partyroom/unregister-me-from-queue/index.ts
@@ -1,0 +1,1 @@
+export { useUnregisterMeFromQueue } from './api/use-unregister-me-from-queue.mutation';
